ROBOT-1235: Made separate packages for trajectory_analyzer and trajectory_analyzer_gui. To use trajectory_analyzer, recommend to use:
pip install --editable <local/path/to/mecademic/package>

To use trajectory_analyzer_gui, recommend:

pip install --editable <local/path/to/trajectory_analyzer/package>
